feat: move components to have a flat architecture
