+++ src/app/app.component.scss
@@ -1,4 +1,4 @@
-@reference "../styles/override/tailwind.css";
+@reference "../styles/extensions/tailwind.css";
 
 #container {
   @apply bg-gray-100 dark:bg-dark-slate-800;
