+++ src/app/features/details/details.component.html
