+++ src/app/features/details/details.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../../styles/override/tailwind.css";
+@reference "../../../styles/extensions/tailwind.css";
 
 .mission {
   @apply bg-white dark:bg-dark-slate-700;
