+++ src/app/features/details/details.component.ts
