+++ src/app/features/email/email.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../../styles/override/tailwind.css";
+@reference "../../../styles/extensions/tailwind.css";
 
 :host {
   --params: 0.8s ease;
