+++ src/app/features/general/general.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../styles/override/tailwind.css";
+@reference "../../../styles/extensions/tailwind.css";
 
 .container {
   @apply text-gray-600 dark:text-dark-gray-300;
