+++ src/app/features/hobbies/hobbies.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../../styles/override/tailwind.css";
+@reference "../../../styles/extensions/tailwind.css";
 
 .hobby {
   display: flex;
