+++ src/app/features/missions/missions.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../../styles/override/tailwind.css";
+@reference "../../../styles/extensions/tailwind.css";
 
 .timeline {
   display: flex;
