+++ src/app/features/resume/resume.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../styles/override/tailwind.css";
+@reference "../../../styles/extensions/tailwind.css";
 
 :host {
   display: grid;
