+++ src/app/features/summary/summary.component.scss
