+++ src/app/shared/components/avatar/avatar.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../styles/override/tailwind.css";
+@reference "../../../../styles/extensions/tailwind.css";
 
 .avatar {
   @apply outline-gray-300! dark:outline-dark-slate-600;
