+++ src/app/shared/components/dialog/dialog.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../styles/override/tailwind.css";
+@reference "../../../../styles/extensions/tailwind.css";
 
 .cv-dialog-mask {
   @apply fixed inset-0 flex items-center justify-center;
