+++ src/app/shared/components/menu/menu.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../styles/override/tailwind.css";
+@reference "../../../../styles/extensions/tailwind.css";
 
 .cv-menu {
   @apply bg-gray-25 dark:bg-dark-slate-700 border border-gray-300 dark:border-dark-slate-600 rounded-sm shadow-lg min-w-[200px] z-50;
