+++ src/app/shared/components/progress-bar/progress-bar.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../styles/override/tailwind.css";
+@reference "../../../../styles/extensions/tailwind.css";
 
 :host {
   flex: 2;
