+++ src/app/shared/components/scroll-top/scroll-top.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../styles/override/tailwind.css";
+@reference "../../../../styles/extensions/tailwind.css";
 
 .scroll-top {
   @apply bg-slate-800 dark:bg-dark-gray-300-light;
