+++ src/app/shared/components/spinner/spinner.component.scss
@@ -1,4 +1,4 @@
-@reference "../../../../styles/override/tailwind.css";
+@reference "../../../../styles/extensions/tailwind.css";
 
 .spinner {
   position: relative;
