+++ src/styles/extensions/ngx-markdown.scss
