+++ src/styles/globals/boilerplate.scss
@@ -1,4 +1,4 @@
-@reference "./override/tailwind.css";
+@reference "./extensions/tailwind.css";
 
 *,
 *::before,
