+++ src/styles/styles.scss
@@ -1,3 +1,3 @@
 @use 'globals/index' as globals;
 @use 'helpers/index' as helpers;
-@use 'override/index' as override;
+@use 'extensions/index' as extensions;
